e2e: move non-test case code to a separate file
